+++ Strange_Lovers/EncoderPart/SpeedEncoder.h
@@ -13,7 +13,7 @@ typedef struct{
 }SpeedEncoder;
 
 
-int SpeedEncoder_get_speed(SpeedEncoder *this_SpeedEncoder);
+float SpeedEncoder_get_speed(SpeedEncoder *this_SpeedEncoder);
 void SpeedEncoder_calc_speed(SpeedEncoder *this_SpeedEncoder);
 void SpeedEncoder_init(SpeedEncoder *this_SpeedEncoder);
